Renamed script package to engine
