+++ lib/engine/arsenic.go
@@ -1,4 +1,4 @@
-package script
+package engine
 
 import (
 	"context"
