+++ lib/engine/filepath.go
@@ -1,4 +1,4 @@
-package script
+package engine
 
 import (
 	"path/filepath"
