+++ lib/engine/script.go
@@ -1,4 +1,4 @@
-package script
+package engine
 
 import (
 	"context"
@@ -8,39 +8,39 @@ import (
 	"github.com/d5/tengo/v2"
 )
 
-var engineModule *EngineModule = &EngineModule{}
+var scriptModule *ScriptModule = &ScriptModule{}
 
-type EngineModule struct {
+type ScriptModule struct {
 	moduleMap  map[string]tengo.Object
 	stopScript context.CancelFunc
 }
 
-func (m *EngineModule) ModuleMap() map[string]tengo.Object {
+func (m *ScriptModule) ModuleMap() map[string]tengo.Object {
 	if m.moduleMap == nil {
 		m.moduleMap = map[string]tengo.Object{
-			"stop": &tengo.UserFunction{Name: "stop", Value: m.stop},
+			"stop": &tengo.UserFunction{Name: "stop", Value: m.tengoStop},
 		}
 	}
 	return m.moduleMap
 }
 
-func (m *EngineModule) stop(args ...tengo.Object) (tengo.Object, error) {
+func (m *ScriptModule) tengoStop(args ...tengo.Object) (tengo.Object, error) {
 	message := ""
 	if len(args) == 1 {
 		message, _ = tengo.ToString(args[0])
 	}
 
-	stopScript(message)
+	m.stop(message)
 	return nil, nil
 }
 
-func stopScript(args ...string) {
+func (m *ScriptModule) stop(args ...string) {
 	if len(args) == 1 {
 		fmt.Println(args[0])
 	}
 
 	go func() {
-		engineModule.stopScript()
+		m.stopScript()
 	}()
 	time.Sleep(1 * time.Millisecond)
 }
